extractor/requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
beautifulsoup4>=4.12.0
PyYAML>=6.0
